+++ include/unity/scopes/ActivationQueryBase.h
@@ -70,6 +70,15 @@ public:
     ActivationQueryBase(Result const& result, ActionMetadata const& metadata, std::string const& widget_id, std::string const& action_id);
 
     /**
+     \brief Create ActivationQueryBase.
+
+     \param update_result The result received by ScopeBase::activate_result_action().
+     \param metadata The metadata received by ScopeBase::activate_result_action().
+     \param action_id The action identifier received by ScopeBase::activate_result_action().
+    */
+    ActivationQueryBase(Result const& update_result, ActionMetadata const& metadata, std::string const& action_id);
+
+    /**
      \brief Called when the originator of the activation request cancelled it.
      */
     virtual void cancelled() override;