+++ include/unity/scopes/ActivationResponse.h
@@ -20,6 +20,7 @@
 
 #include <unity/scopes/Variant.h>
 #include <unity/scopes/CannedQuery.h>
+#include <unity/scopes/Result.h>
 #include <memory>
 
 namespace unity
@@ -41,23 +42,25 @@ class ActivationResponse final
 {
 public:
     /**
-     \brief Status of a unity::scopes::ScopeBase::activate or unity::scopes::ScopeBase::perform_action request.
+     \brief Status of a unity::scopes::ScopeBase::activate, unity::scopes::ScopeBase::perform_action,
+     or unity::scopes::ScopeBase::activate_result_action request.
      */
     enum Status
     {
-        NotHandled,  /**< Activation of this result wasn't handled by the scope */
-        ShowDash,    /**< Activation of this result was handled, show the Dash */
-        HideDash,    /**< Activation of this result was handled, hide the Dash */
-        ShowPreview, /**< Preview should be requested for this result */
-        PerformQuery /**< Perform new search. This state is implied if creating ActivationResponse with CannedQuery object and is invalid otherwise */
+        NotHandled,   /**< Activation of this result wasn't handled by the scope */
+        ShowDash,     /**< Activation of this result was handled, show the Dash */
+        HideDash,     /**< Activation of this result was handled, hide the Dash */
+        ShowPreview,  /**< Preview should be requested for this result */
+        PerformQuery, /**< Perform new search. This state is implied if creating ActivationResponse with CannedQuery object and is invalid otherwise */
+        UpdateResult, /**< Update the result. This state is implied if creating ActivationResponse with Result object and is invalid otherwise */
     };
 
     /**
     \brief Creates ActivationResponse with given status.
     \param status The activation status.
-    \throws unity::InvalidArgumentException if status is Status::PerformQuery. To
-    create an ActivationResponse of that type, use the ActivationResponse(CannedQuery const&)
-    constructor.
+    \throws unity::InvalidArgumentException if status is Status::PerformQuery or Status::Update.
+    To create an ActivationResponse for one of these types, use the ActivationResponse(CannedQuery const&)
+    or ActivationResponse(Result const&) constructors.
     */
     ActivationResponse(Status status);
 
@@ -66,6 +69,13 @@ public:
     \param query The search query to be executed by the client.
      */
     ActivationResponse(CannedQuery const& query);
+
+    /**
+    \brief Creates an ActivationResponse with status Status::UpdateResult and an updated result that
+    replaces the original result of the action.
+    \param updated_result The updated result to replace the original result of the action.
+     */
+    ActivationResponse(Result const& updated_result);
 
     /**@name Copy and assignment
     Copy and assignment operators (move and non-move versions) have the usual value semantics.
@@ -88,13 +98,13 @@ public:
      \brief Attach arbitrary data to this response.
 
      The attached data is sent back to the scope if the status of this response is Status::ShowPreview.
-     \param data arbitrary value attached to response
+     \param data An arbitrary value attached to the response.
      */
     void set_scope_data(Variant const& data);
 
     /**
      \brief Get data attached to this response object.
-     \return The data attached to response.
+     \return The data attached to the response.
      */
     Variant scope_data() const;
 
@@ -105,6 +115,14 @@ public:
      \return The query to be executed by the client.
     */
     CannedQuery query() const;
+
+    /**
+     \brief The updated result if status is Status::UpdateResult.
+
+     \throws unity::LogicException if the status of this ActivationResponse is anything other than Status::UpdateResult.
+     \return The result to be displayed instead of the original result.
+    */
+    Result updated_result() const;
 
     /// @cond
     VariantMap serialize() const;