+++ test/gtest/scopes/ActivationResponse/ActivationResponse_test.cpp
@@ -19,6 +19,8 @@
 #include <unity/scopes/ActivationResponse.h>
 #include <unity/scopes/CannedQuery.h>
 #include <unity/scopes/internal/ActivationResponseImpl.h>
+#include <unity/scopes/internal/ResultImpl.h>
+#include <unity/scopes/internal/RuntimeImpl.h>
 #include <unity/UnityExceptions.h>
 #include <gtest/gtest.h>
 
@@ -53,12 +55,35 @@ TEST(ActivationResponse, basic)
         ActivationResponse resp(query);
         EXPECT_EQ(ActivationResponse::Status::PerformQuery, resp.status());
     }
+    {
+        VariantMap vm;
+        VariantMap attrs;
+        attrs["uri"] = "foo";
+        vm["internal"] = Variant(VariantMap());
+        vm["attrs"] = Variant(attrs);
+        auto const res = ResultImpl::create_result(vm);
+
+        ActivationResponse resp(res);
+        EXPECT_EQ(ActivationResponse::Status::UpdateResult, resp.status());
+        EXPECT_EQ(res.uri(), resp.updated_result().uri());
+    }
+
 
     // Search only allowed with CannedQuery
     {
         try
         {
             ActivationResponse resp(ActivationResponse::Status::PerformQuery);
+            FAIL();
+        }
+        catch (unity::InvalidArgumentException const&) {}
+    }
+
+    // UpdateResult only allowed with Result
+    {
+        try
+        {
+            ActivationResponse resp(ActivationResponse::Status::UpdateResult);
             FAIL();
         }
         catch (unity::InvalidArgumentException const&) {}
@@ -85,16 +110,32 @@ TEST(ActivationResponse, serialize)
         EXPECT_EQ(ActivationResponse::Status::PerformQuery, static_cast<ActivationResponse::Status>(var["status"].get_int()));
         EXPECT_EQ("scope-foo", var["query"].get_dict()["scope"].get_string());
     }
+    {
+        VariantMap vm;
+        VariantMap attrs;
+        attrs["uri"] = Variant("foo");
+        vm["internal"] = Variant(VariantMap());
+        vm["attrs"] = Variant(attrs);
+        auto const res = ResultImpl::create_result(vm);
+
+        ActivationResponse resp(res);
+        auto var = resp.serialize();
+        EXPECT_EQ(ActivationResponse::Status::UpdateResult, static_cast<ActivationResponse::Status>(var["status"].get_int()));
+        EXPECT_EQ("foo", var["updated_result"].get_dict()["attrs"].get_dict()["uri"].get_string());
+    }
 }
 
 TEST(ActivationResponse, deserialize)
 {
+    // just to make ResultImpl::set_runtime() happy, runtime must not be null.
+    auto runtime = internal::RuntimeImpl::create("fooscope", TEST_DIR "/Runtime.ini");
+
     // invalid variant
     {
         VariantMap var;
         try
         {
-            internal::ActivationResponseImpl res(var);
+            internal::ActivationResponseImpl res(var, runtime.get());
             FAIL();
         }
         catch (unity::LogicException const &e)
@@ -108,7 +149,7 @@ TEST(ActivationResponse, deserialize)
         var["scope_data"] = VariantMap();
         try
         {
-            internal::ActivationResponseImpl::create(var);
+            internal::ActivationResponseImpl::create(var, runtime.get());
             FAIL();
         }
         catch (unity::LogicException const &e)
@@ -125,7 +166,7 @@ TEST(ActivationResponse, deserialize)
         var["status"] = static_cast<int>(ActivationResponse::Status::PerformQuery);
         try
         {
-            internal::ActivationResponseImpl res(var);
+            internal::ActivationResponseImpl res(var, runtime.get());
             FAIL();
         }
         catch (unity::LogicException const &e) {}
@@ -140,7 +181,7 @@ TEST(ActivationResponse, deserialize)
         var["status"] = static_cast<int>(ActivationResponse::Status::HideDash);
         try
         {
-            internal::ActivationResponseImpl res(var);
+            internal::ActivationResponseImpl res(var, runtime.get());
             EXPECT_EQ("bar", res.scope_data().get_dict()["foo"].get_string());
             EXPECT_EQ(ActivationResponse::Status::HideDash, res.status());
         }
@@ -159,7 +200,7 @@ TEST(ActivationResponse, deserialize)
         var["query"] = query.serialize();
         try
         {
-            internal::ActivationResponseImpl res(var);
+            internal::ActivationResponseImpl res(var, runtime.get());
             EXPECT_EQ(ActivationResponse::Status::PerformQuery, res.status());
             EXPECT_EQ("scope-foo", res.query().scope_id());
         }
@@ -176,7 +217,7 @@ TEST(ActivationResponse, deserialize)
         var["status"] = static_cast<int>(ActivationResponse::Status::HideDash);
         try
         {
-            internal::ActivationResponseImpl res(var);
+            internal::ActivationResponseImpl res(var, runtime.get());
             EXPECT_EQ("foobar", res.scope_data().get_string());
             EXPECT_EQ(ActivationResponse::Status::HideDash, res.status());
         }